fix Augury integration
fixes Augury errors in projects that include these components, and
removes the `forwardRef`s that abounded by adding a service to the
components to deal with menu & dropdown state.

The service is instantiated once for each dropdown, so won't interfere
with multiple instantiations of the dropdown.

Tests changed just enough to use the new object structure in the state
(plus the test component module now imports the actual dropdown module,
not just all of the components in it)